Implement automatic table generation of default prefixes
