+++ crates/sparql_queries/Cargo.toml
@@ -18,6 +18,7 @@
 [dependencies]
     grapher={workspace=true}
     indexmap="2.14"
+    leptos.workspace=true
     log.workspace=true
     regex="1.12.3"
     vowlgrapher-util={path="../util"}
