+++ crates/sparql_queries/src/assembly.rs
@@ -1,26 +1,15 @@
 pub mod normalization;
+pub mod prefix;
 pub mod query_regex;
 
 use crate::assembly::normalization::QueryNormalizer;
+use crate::assembly::prefix::DEFAULT_PREFIXES;
 use crate::errors::QueryAssemblyError;
 use crate::prelude::GENERAL_SNIPPETS;
 use crate::snippets::SparqlSnippet;
 use crate::snippets::void::VOID;
 use grapher::prelude::ElementType;
 use std::collections::HashMap;
-
-// TODO: Remove when automatic prefix fetching is implemented.
-pub const DEFAULT_PREFIXES: [&str; 9] = [
-    "example: <http://www.example.com/iri#>",
-    "vowlgrapher: <https://purl.org/vowlgrapher>",
-    "owl: <http://www.w3.org/2002/07/owl#>",
-    "rdfs: <http://www.w3.org/2000/01/rdf-schema#>",
-    "rdf: <http://www.w3.org/1999/02/22-rdf-syntax-ns#>",
-    "xsd: <http://www.w3.org/2001/XMLSchema#>",
-    "xml: <http://www.w3.org/XML/1998/namespace>",
-    "dc: <http://purl.org/dc/elements/1.1/>",
-    "dcterms: <http://purl.org/dc/terms/>",
-];
 
 /// Compiles snippets of SPARQL code into full-fledged SPARQL queries.
 pub struct QueryAssembler;
